+++ src/traits.rs
@@ -2,6 +2,7 @@
 
 mod encryption;
 pub(crate) mod keys;
+pub(crate) mod modular;
 mod padding;
 
 pub use encryption::{Decryptor, EncryptingKeypair, RandomizedDecryptor, RandomizedEncryptor};
@@ -9,4 +10,5 @@ pub use encryption::{Decryptor, EncryptingKeypair, RandomizedDecryptor, Randomiz
 pub use keys::{PrivateKeyParts, PublicKeyParts};
 #[cfg(not(feature = "private-key"))]
 pub use keys::{PublicKeyParts};
+pub use modular::UnsignedModularInt;
 pub use padding::{PaddingScheme, SignatureScheme};
